utf-8 comment, author notice, moved colour space converters
