+++ EVM/mode.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+# Volker Göhler 2019,2020
 import enum
